requirements.txt
streamlit
groq
python-dotenv
streamlit-authenticator
pyyaml
bcrypt
fpdf2